明确定义，rename construct->args
