+++ src/manimgeo/components/angle/__init__.py
@@ -1,3 +1,3 @@
 from .angle import Angle
 from .adapter import AngleAdapter
-from .construct import AngleConstructArgsList
+from .args import AngleConstructArgsList
