+++ src/manimgeo/components/circle/__init__.py
@@ -1,3 +1,3 @@
 from .circle import Circle
 from .adapter import CircleAdapter
-from .construct import CircleConstructArgsList
+from .args import CircleConstructArgsList
