+++ src/manimgeo/components/line/__init__.py
@@ -1,3 +1,3 @@
 from .line import Line, LineSegment, Ray, InfinityLine
 from .adapter import LineAdapter
-from .construct import LineConstructArgsList
+from .args import LineConstructArgsList
