+++ src/manimgeo/components/vector/__init__.py
@@ -1,3 +1,3 @@
 from .vector import Vector
 from .adapter import VectorAdapter
-from .construct import VectorConstructArgsList
+from .args import VectorConstructArgsList
